djangotest/requirements.txt
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.4.1
